restaurant-finder-api/pyproject.toml
[project]
name = "restaurant-finder-api"
version = "0.1.0"
description = "Restaurant Finder AI Agent API"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "bedrock-agentcore>=1.2.0",
    "bedrock-agentcore-starter-toolkit>=0.2.8",
    "boto3>=1.42.37",
    "langchain>=0.3.0",
    "langchain-aws>=0.2.0",
    "langchain-mcp-adapters>=0.1.11",
    "langgraph>=1.0.7",
    "langgraph-checkpoint-aws>=1.0.4",
    "loguru>=0.7.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.0",
    # Observability dependencies
    "aws-opentelemetry-distro>=0.12.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
]
